some fixex
