+++ js/slider-effects.js
@@ -105,24 +105,4 @@ const onChangeEffectValue = (handlersValue) => {
   effectInputValue.value = value;
 };
 
-const uiSlider = noUiSlider.create(effectLevelSlider, {
-  range: { min: 0, max: 1 },
-  start: 1,
-  step: 0.1,
-  connect: 'lower',
-  format: {
-    to: function (value) {
-      if (Number.isInteger(value)) {
-        return value.toFixed(0);
-      }
-      return value.toFixed(1);
-    },
-    from: function (value) {
-      return parseFloat(value);
-    },
-  },
-});
-
-uiSlider.on('update', onChangeEffectValue);
-
 export { onChangeImageEffect, onChangeEffectValue };
