+++ js/upload-image.js
@@ -1,7 +1,7 @@
 import { isEscapeKey } from './utils.js';
 import { openMessagePopup } from './message-popup.js';
 import { addScaleHandler, removeScaleHandler } from './scale-editor.js';
-import { onChangeImageEffect } from './slider-effects.js';
+import { onChangeImageEffect, onChangeEffectValue } from './slider-effects.js';
 import { uploadFormValidate } from './form.js';
 
 const IMAGE_SCALE = 100;
@@ -25,7 +25,7 @@ const effectLevelSlider = imageUpload.querySelector('.effect-level__slider');
 const imageEffects = document.querySelector('.img-upload__effects');
 const imageLevelEffect = document.querySelector('.img-upload__effect-level');
 
-//Все данные, введённые в форму, и контрол фильтра приходят в исходное состояние:
+// Все данные, введённые в форму, и контрол фильтра приходят в исходное состояние:
 
 const clearUserEnterData = () => {
   scaleControl.value = `${IMAGE_SCALE}%`;
@@ -36,7 +36,8 @@ const clearUserEnterData = () => {
   imageUploadPreview.src = '';
 };
 
-//Форма редактирования изображения закрывается
+// Форма редактирования изображения закрывается
+
 const closeEditImagePopup = () => {
   imageUploadPopup.classList.add('hidden');
   document.body.classList.remove('modal-open');
@@ -50,7 +51,7 @@ const closeEditImagePopup = () => {
   clearUserEnterData();
 };
 
-//Форма редактирования изображения открывается
+// Форма редактирования изображения открывается
 
 function openEditImagePopup() {
   const file = this.files[0];
@@ -69,6 +70,26 @@ function openEditImagePopup() {
   addScaleHandler();
 
   document.addEventListener('keydown', onEditPopupEsc);
+
+  const uiSlider = noUiSlider.create(effectLevelSlider, {
+    range: { min: 0, max: 1 },
+    start: 1,
+    step: 0.1,
+    connect: 'lower',
+    format: {
+      to: function (value) {
+        if (Number.isInteger(value)) {
+          return value.toFixed(0);
+        }
+        return value.toFixed(1);
+      },
+      from: function (value) {
+        return parseFloat(value);
+      },
+    },
+  });
+
+  uiSlider.on('update', onChangeEffectValue);
 
   imageEffects.addEventListener('change', onChangeImageEffect);
